Rename mpprest to rest (#7804)
