+++ server/src/main/java/org/apache/iotdb/db/protocol/rest/impl/PingApiServiceImpl.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.iotdb.db.protocol.mpprest.impl;
+package org.apache.iotdb.db.protocol.rest.impl;
 
 import org.apache.iotdb.commons.service.ThriftService;
-import org.apache.iotdb.db.protocol.mpprest.PingApiService;
-import org.apache.iotdb.db.protocol.mpprest.model.ExecutionStatus;
+import org.apache.iotdb.db.protocol.rest.PingApiService;
+import org.apache.iotdb.db.protocol.rest.model.ExecutionStatus;
 import org.apache.iotdb.db.service.RPCService;
 import org.apache.iotdb.rpc.TSStatusCode;
